По замечаниям PR #2: выделен класс утилит, доработаны методы сохранения в историю и тесты.
